backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.0
celery==5.3.6
redis==5.0.3
httpx==0.27.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
dbt-core==1.7.0
dbt-postgres==1.7.0
databricks-sdk==0.28.0
